+++ packages/react-auth-core/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.5.3](https://github.com/ttoss/ttoss/compare/@ttoss/react-auth-core@0.5.2...@ttoss/react-auth-core@0.5.3) (2026-06-22)
+
+**Note:** Version bump only for package @ttoss/react-auth-core
+
+## [0.5.2](https://github.com/ttoss/ttoss/compare/@ttoss/react-auth-core@0.5.1...@ttoss/react-auth-core@0.5.2) (2026-06-22)
+
+**Note:** Version bump only for package @ttoss/react-auth-core
+
 ## [0.5.1](https://github.com/ttoss/ttoss/compare/@ttoss/react-auth-core@0.5.0...@ttoss/react-auth-core@0.5.1) (2026-06-19)
 
 **Note:** Version bump only for package @ttoss/react-auth-core